+++ src/adapters/index.ts
@@ -3,3 +3,5 @@
  */
 export { QQAdapter } from "./qq/index";
 export { DiscordAdapter } from "./discord";
+export { MultiAdapter } from "./multi";
+export { createPlatformAdapters } from "./factory";